\name{exportedPlus}
\alias{exportedPlus}
\alias{privatePlus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Add two numbers together }
\description{
  Simple function for testing RtTests package
}
\usage{
privatePlus(x, y)
exportedPlus(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ one argument }
  \item{y}{ the other }
}
\details{
  Just returns the result of \code{x + y}.
}
\value{
  The result of \code{x + y}.
}
\author{ Tony Plate }
\examples{
1 + 2
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }
